fixed up the tailwind setup and moved current pages to a new folder
