--- frontend/postcss.config.js
@@ -1,10 +1,0 @@
-// postcss.config.js
-import tailwindcss from '@tailwindcss/postcss';
-import autoprefixer from 'autoprefixer';
-
-export default {
-  plugins: [
-    tailwindcss,
-    autoprefixer
-  ],
-}
