+++ frontend/src/pages/CreateAccountPage.jsx
